+++ CHANGELOG.md
@@ -15,6 +15,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Changed script structure, deleting function to avoid declaring parameters twice.
 
+### Added
+
+- Added default type.ps1xml file to project.
+
 ## 1.1.0 - 2021-03-02
 
 ### Added